backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Deprecated==1.2.13
django-debug-toolbar==3.6.0
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
Faker==14.2.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
packaging==21.3
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.8.6
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
wrapt==1.14.1
zipp==3.8.1
